+++ src/App.js
@@ -1,137 +1,12 @@
-import React, { Component } from "react";
-import DisplayResult from "./component/DisplayResult";
-import CalculatorButton from "./component/CalculatorButton";
+import { Component } from "react";
+import Calculator from "./component/Calculator";
 
 export default class App extends Component {
-  state = {
-    firstNumber: 0,
-    operator: null,
-    secondNumber: 0,
-    isFirstNumber: true,
-    result: 0,
-  };
-
-  componentDidMount() {
-    const prevValue = localStorage.getItem("prevValue") || 0;
-    this.setFirstNumber(Number(prevValue));
-    this.setResult(Number(prevValue));
-
-    window.addEventListener("beforeunload", function (e) {
-      e.preventDefault();
-      e.returnValue = "";
-    });
-  }
-
-  initState = () => {
-    this.setState({
-      firstNumber: 0,
-      operator: null,
-      secondNumber: 0,
-      isFirstNumber: true,
-      result: 0,
-    });
-  };
-
-  setFirstNumber = (number) => {
-    this.setState(
-      {
-        firstNumber: Number(number),
-      },
-      localStorage.setItem("prevValue", number)
-    );
-  };
-
-  setOperator = (operator) => {
-    this.setState({ operator });
-  };
-
-  setSecondNumber = (number) => {
-    this.setState(
-      {
-        secondNumber: Number(number),
-      },
-      localStorage.setItem("prevValue", number)
-    );
-  };
-
-  setResult = (result) => {
-    this.setState({ result });
-  };
-
-  setIsFirstNumber = (isFirstNumber) => {
-    this.setState({ isFirstNumber });
-  };
-
-  calculate = () => {
-    const res = (() => {
-      switch (this.state.operator) {
-        case "+":
-          return this.add();
-        case "-":
-          return this.sub();
-        case "X":
-          return this.multiple();
-        case "/":
-          return this.divide();
-        default:
-          throw new Error("존재하지 않는 연산자입니다.");
-      }
-    })();
-
-    this.setState({ result: res }, () => {
-      this.initState();
-
-      if (res === Infinity || isNaN(res)) {
-        this.setFirstNumber(0);
-        this.setResult("오류");
-        localStorage.setItem("prevValue", "오류");
-        return;
-      }
-
-      this.setFirstNumber(res);
-      this.setResult(res);
-      localStorage.setItem("prevValue", res);
-    });
-  };
-
-  add() {
-    return this.state.firstNumber + this.state.secondNumber;
-  }
-
-  sub() {
-    return this.state.firstNumber - this.state.secondNumber;
-  }
-
-  divide() {
-    return Math.floor(this.state.firstNumber / this.state.secondNumber);
-  }
-
-  multiple() {
-    return this.state.firstNumber * this.state.secondNumber;
-  }
-
   render() {
     return (
-      <>
-        <div id="app">
-          <div className="calculator">
-            <DisplayResult result={this.state.result} />
-            <CalculatorButton
-              setFirstNumber={this.setFirstNumber}
-              setOperator={this.setOperator}
-              setSecondNumber={this.setSecondNumber}
-              setIsFirstNumber={this.setIsFirstNumber}
-              calculate={this.calculate}
-              setResult={this.setResult}
-              result={this.state.result}
-              initState={this.initState}
-              isFirstNumber={this.state.isFirstNumber}
-              firstNumber={this.state.firstNumber}
-              secondNumber={this.state.secondNumber}
-            />
-          </div>
-        </div>
-      </>
+      <div id="app">
+        <Calculator />
+      </div>
     );
   }
 }
