+++ src/component/DisplayResult.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import { Component } from "react";
 
 export default class DisplayResult extends Component {
   render() {
